packages/python/goldenpipe/pyproject.toml
[project]
name = "goldenpipe"
version = "1.0.0"
description = "Pluggable pipeline framework for data quality workflows"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "polars>=1.0",
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
check = ["goldencheck>=0.5.0"]
flow = ["goldenflow>=0.1.0"]
match = ["goldenmatch>=1.2.0"]
golden-suite = ["goldencheck>=0.5.0", "goldenflow>=0.1.0", "goldenmatch>=1.2.0"]
tui = ["textual>=1.0"]
api = ["fastapi>=0.110", "uvicorn>=0.29"]
mcp = ["mcp>=1.0"]
agent = ["aiohttp>=3.9"]
all = ["goldenpipe[golden-suite,tui,api,mcp,agent]"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "pytest-aiohttp>=1.0", "httpx>=0.27"]

[project.scripts]
goldenpipe = "goldenpipe.cli.main:app"

[project.entry-points."goldenpipe.stages"]
"goldencheck.scan" = "goldenpipe.adapters.check:ScanStage"
"goldenflow.transform" = "goldenpipe.adapters.flow:TransformStage"
"goldenmatch.dedupe" = "goldenpipe.adapters.match:DedupeStage"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
